service-payment/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
aio-pika
psycopg2-binary
sqlalchemy>=1.4.0
